requirements.txt
# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
pydantic-settings>=2.5.0

# Computer Vision
opencv-python>=4.10.0

# Deep Learning & ML
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.3.0
numpy>=1.24.0,<2.0
scipy>=1.10.0
scikit-image>=0.20.0

# Image Processing
Pillow>=10.0.0
matplotlib>=3.7.0

# Data Handling
pandas>=2.0.0

# Utilities
python-multipart>=0.0.6
python-dotenv>=1.0.0
tqdm>=4.66.0

# Authentication
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0

# Logging
python-json-logger>=2.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Code Quality
black>=23.0.0
flake8>=6.1.0
isort>=5.12.0
mypy>=1.5.0

# Similarity Search
faiss-cpu>=1.7.4

# MLOps
mlflow>=2.14.0

# Monitoring
prometheus-client>=0.18.0
